Backend header generation fixes
Added stdarg.h include (for va_list), and stddef.h (for size_t) back into the header.
Added templates for SDL2/3 backends and removed the old SDL ones.
Fixed a bug affecting backend generation where if an extern "C" block was immediately followed by a preprocessor
directive, it incorrectly omitted the brackets from the extern thinking that it was a single statement.